% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_pub.r
\docType{data}
\name{geom_pub}
\alias{geom_pub}
\title{geom_pub}
\format{An object of class \code{list} of length 3.}
\usage{
geom_pub
}
\description{
geom values to be used with published figures
}
\keyword{datasets}
