% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geom_defaults.r
\name{get_geom_defaults}
\alias{get_geom_defaults}
\title{get_geom_defaults}
\usage{
get_geom_defaults(geom)
}
\arguments{
\item{geom}{a string for any of the ggplot geoms}
}
\value{
list of defaults
}
\description{
Gets a list of defaults for any given ggplot geom
}
