% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpup.r
\name{ggpub}
\alias{ggpub}
\title{ggpub}
\usage{
ggpub(filename, width, height, plot = last_plot(), geoms = geom_pub,
  theme = theme_pub)
}
\arguments{
\item{filename}{string for the file to export (no expension)}

\item{width}{the width in mm}

\item{height}{the height in mm}

\item{plot}{the plot to export (last plot if NONE)}

\item{geoms}{list of geoms to change in the form list(geom1 = list(prop1 = val1), etc).}

\item{theme}{the theme to use}
}
\value{
NONE
}
\description{
Exports a PDF of a plot with sane defaults for publications (eg
much smaller text and dimensions).
}
