% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luminex.r
\name{import_luminex}
\alias{import_luminex}
\title{import_luminex}
\usage{
import_luminex(filepath)
}
\arguments{
\item{filepath}{string for the file to import}
}
\value{
tidy tibble with columns for sample number, well, analyte, and concentration. Note that standard curves are ignored
}
\description{
Imports luminex data from the multi-sheet output given
from the biorad luminex machine. All sheets are combined
into one output. The input is expected to be a .xlsx file
}
