% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_sheets.R
\name{read_excel_sheets}
\alias{read_excel_sheets}
\title{read_excel_sheets}
\usage{
read_excel_sheets(filepath, ...)
}
\arguments{
\item{filepath}{to the file to be imported.}

\item{...}{args to supply to read_excel}
}
\value{
a list of tibbles with each index as the name of the imported sheet.
}
\description{
reads all sheets from an excel file
}
