% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pub.r
\docType{data}
\name{theme_pub}
\alias{theme_pub}
\title{theme_pub}
\format{An object of class \code{theme} (inherits from \code{gg}) of length 57.}
\usage{
theme_pub
}
\description{
theme for use in published figures
}
\keyword{datasets}
